requirements.txt
python-telegram-bot>=21.0
openpyxl>=3.1.0
pandas>=2.0.0
groq>=0.4.0
